requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
coreapi==2.3.3
python-dotenv==1.0.0
requests==2.31.0
django-ratelimit==4.1.0
